refactor : interface 수정 - createdAt: Data타입으로 변경 - timestamps적용
