--- src/interfaces/login.ts
@@ -1,6 +1,0 @@
-import mongoose, { Document } from 'mongoose';
-
-export interface ILogin extends Document {
-  userId: mongoose.Types.ObjectId;
-  loginTime: number;
-}
