+++ src/schemas/Like.ts
@@ -1,16 +1,13 @@
 import mongoose, { Schema } from 'mongoose';
-import { createdDate } from '../utils';
 import { IBase } from '../interfaces';
 
-const LikeSchema = new Schema<IBase>({
-  contentId: { type: Schema.Types.ObjectId, require: true, ref: 'Content' },
-  userId: { type: Schema.Types.ObjectId, require: true, ref: 'User' },
-  createdAt: {
-    type: Number,
-    require: true,
-    default: createdDate,
+const LikeSchema = new Schema<IBase>(
+  {
+    contentId: { type: Schema.Types.ObjectId, require: true, ref: 'Content' },
+    userId: { type: Schema.Types.ObjectId, require: true, ref: 'User' },
   },
-});
+  { timestamps: true },
+);
 
 LikeSchema.index({ contentId: 1, userId: 1 }, { unique: true });
