+++ src/schemas/Login.ts
@@ -1,18 +1,15 @@
 import mongoose, { Schema } from 'mongoose';
-import { createdDate } from '../utils';
-import { ILogin } from '../interfaces';
+import { ILoginRecord } from '../interfaces';
 
-const LoginSchema = new Schema<ILogin>({
-  userId: {
-    type: Schema.Types.ObjectId,
-    ref: 'User',
-    required: true,
+const LoginSchema = new Schema<ILoginRecord>(
+  {
+    userId: {
+      type: Schema.Types.ObjectId,
+      ref: 'User',
+      required: true,
+    },
   },
-  loginTime: {
-    type: Number,
-    required: true,
-    default: createdDate,
-  },
-});
+  { timestamps: true },
+);
 
-export default mongoose.model<ILogin>('Login', LoginSchema);
+export default mongoose.model<ILoginRecord>('Login', LoginSchema);
