+++ src/schemas/Share.ts
@@ -1,16 +1,13 @@
 import { IShare } from '../interfaces';
-import { createdDate } from '../utils';
 import mongoose, { Schema } from 'mongoose';
 
-const ShareSchema = new Schema<IShare>({
-  contentId: { type: Schema.Types.ObjectId, required: true, ref: 'Content' },
-  userId: { type: Schema.Types.ObjectId, ref: 'Member' },
-  createdAt: {
-    type: Number,
-    require: true,
-    default: createdDate,
+const ShareSchema = new Schema<IShare>(
+  {
+    contentId: { type: Schema.Types.ObjectId, required: true, ref: 'Content' },
+    userId: { type: Schema.Types.ObjectId, ref: 'Member' },
+    type: { type: String, required: true, enum: ['Kakao', 'Link'] },
   },
-  type: { type: String, required: true, enum: ['Kakao', 'Link'] },
-});
+  { timestamps: true },
+);
 
 export default mongoose.model<IShare>('Share', ShareSchema);
