+++ src/schemas/User.ts
@@ -1,26 +1,23 @@
 import mongoose, { Schema } from 'mongoose';
-import { createdDate } from '../utils';
 import { IUser } from '../interfaces/user';
 
-const UserSchema = new Schema<IUser>({
-  kakaoId: {
-    type: Number,
-    required: true,
-    unique: true,
+const UserSchema = new Schema<IUser>(
+  {
+    kakaoId: {
+      type: Number,
+      required: true,
+      unique: true,
+    },
+    name: { type: String, required: true },
+    thumbnail: { type: String, required: true },
+    role: {
+      type: String,
+      required: true,
+      enum: ['User', 'Admin', 'Creator'],
+      default: 'User',
+    },
   },
-  name: { type: String, required: true },
-  thumbnail: { type: String, required: true },
-  createdAt: {
-    type: Number,
-    required: true,
-    default: createdDate,
-  },
-  role: {
-    type: String,
-    required: true,
-    enum: ['User', 'Admin', 'Creator'],
-    default: 'User',
-  },
-});
+  { timestamps: true },
+);
 
 export default mongoose.model<IUser>('User', UserSchema);
