+++ src/schemas/UserResult.ts
@@ -1,16 +1,13 @@
 import mongoose, { Schema } from 'mongoose';
-import { createdDate } from '../utils';
 import { IUserResult } from '../interfaces';
 
-const UserResultSchema = new Schema<IUserResult>({
-  contentId: { type: Schema.Types.ObjectId, required: true, ref: 'Content' },
-  resultId: { type: Schema.Types.ObjectId, required: true, ref: 'Result' },
-  userId: { type: Schema.Types.ObjectId, required: true, ref: 'Member' },
-  createdAt: {
-    type: Number,
-    required: true,
-    default: createdDate,
+const UserResultSchema = new Schema<IUserResult>(
+  {
+    contentId: { type: Schema.Types.ObjectId, required: true, ref: 'Content' },
+    resultId: { type: Schema.Types.ObjectId, required: true, ref: 'Result' },
+    userId: { type: Schema.Types.ObjectId, required: true, ref: 'Member' },
   },
-});
+  { timestamps: true },
+);
 
 export default mongoose.model<IUserResult>('UserResult', UserResultSchema);
