+++ src/types/index.ts
@@ -2,3 +2,10 @@ export * from './auth';
 
 export type ContentType = 'MBTI';
 export type ShareType = 'Kakao' | 'Link';
+export type Sort = 'asc' | 'desc';
+
+export type PaginationOptions = {
+  size: string;
+  page: string;
+  sort?: Sort;
+};
